Update to match latest upstream
The toolsets in SDK 0.9 have been reorganised which means that the old linaro/carbon branch is no longer functional. However, since carbon is now fully upstreamed there's no point in using this branch anyway.

Update the instructions to use the zephyr mainline instead and update the directory where the shell example is located.