96b-som-carrier: added basic documentation
I have changed the doc layout for the carrier board to
match that of the mezzanine section.

Since the concept in pretty much similar of the carrier
board being a non-standalone accessory to SoM modules, like
mezzanines are to CE boards, it seems justified to keep the
doc structure the same as that of the mezzanine section.

- added basic guide
- changed layout

TODO:
- add mux discription guide

Signed-off-by: Sahaj Sarup <sahaj.sarup@linaro.org>